move CheckRPCStatus to observer.go
