+++ zetaclient/chains/solana/observer/observer.go
@@ -1,13 +1,17 @@
 package observer
 
 import (
+	"context"
+
 	"github.com/gagliardetto/solana-go"
 	"github.com/gagliardetto/solana-go/rpc"
 	"github.com/pkg/errors"
 
+	"github.com/zeta-chain/node/pkg/chains"
 	contracts "github.com/zeta-chain/node/pkg/contracts/solana"
 	"github.com/zeta-chain/node/zetaclient/chains/base"
 	"github.com/zeta-chain/node/zetaclient/chains/interfaces"
+	zetasolrpc "github.com/zeta-chain/node/zetaclient/chains/solana/rpc"
 )
 
 // Observer is the observer for the Solana chain
@@ -81,4 +85,20 @@ func (ob *Observer) IsTxFinalized(nonce uint64) bool {
 	ob.Mu().Lock()
 	defer ob.Mu().Unlock()
 	return ob.finalizedTxResults[ob.OutboundID(nonce)] != nil
+}
+
+// checkRPCStatus checks the RPC status of the Solana chain
+func (ob *Observer) CheckRPCStatus(ctx context.Context) error {
+	// Solana privnet doesn't have RPC 'GetHealth', need to differentiate
+	privnet := ob.Chain().NetworkType == chains.NetworkType_privnet
+
+	// check the RPC status
+	blockTime, err := zetasolrpc.CheckRPCStatus(ctx, ob.solClient, privnet)
+	if err != nil {
+		return errors.Wrap(err, "unable to check rpc status")
+	}
+
+	ob.ReportBlockLatency(blockTime)
+
+	return nil
 }
